feat: running in the web
