+++ .vscode/bevy_gaussian_splatting.code-workspace
@@ -7,5 +7,7 @@
 			"path": "../../bevy"
 		}
 	],
-	"settings": {}
+	"settings": {
+		"liveServer.settings.multiRootWorkspaceName": "bevy_gaussian_splatting"
+	}
 }
