+++ Cargo.toml
@@ -36,8 +36,10 @@ default-run = "viewer"
 default = [
   "io_flexbuffers",
   "io_ply",
+  "morph_particles",
   "sort_radix",
   "sort_rayon",
+  "sort_std",
   "tooling",
   "viewer",
 ]
@@ -48,11 +50,13 @@ io_bincode2 = ["bincode2", "flate2"]
 io_flexbuffers = ["flexbuffers"]
 io_ply = ["ply-rs"]
 
+morph_particles = []
+
 sort_radix = []
 sort_rayon = [
   "rayon",
-  "wasm-bindgen-rayon",
 ]
+sort_std = []
 
 tooling = [
   "byte-unit",
@@ -62,7 +66,6 @@ viewer = [
   "bevy-inspector-egui",
   "bevy_panorbit_camera",
 ]
-
 
 
 [dependencies]
@@ -84,21 +87,18 @@ wgpu = "0.17.1"
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 console_error_panic_hook = "0.1"
 wasm-bindgen = "0.2"
-wasm-bindgen-rayon = { version = "1.0", optional = true }
 
 
 # TODO: use minimal bevy features
 [dependencies.bevy]
 version = "0.12"
-default-features = true
-
-[dev-dependencies.bevy]
-version = "0.12"
-default-features = true
+default-features = false
 features = [
-  'debug_glam_assert',
+  "bevy_asset",
+  "bevy_core_pipeline",
+  "bevy_render",
+  "bevy_winit",
 ]
-
 
 
 [dependencies.web-sys]
@@ -107,6 +107,7 @@ features = [
   'Document',
   'Element',
   'HtmlElement',
+  'Location',
   'Node',
   'Window',
 ]
